v0.53
Made some visual layout changes and hardcoded technical skills into the webpage. Decided that I had more control over the look of the website if I just used a simpler method. No need to be a show off with that pulling json data...its really not that impressive to begin with.